Created Add and Remove Func for MultiMap
Extension to the normal Dictionary. This class can store more than one value for every key.
It keeps a HashSet for every Key value.  Calling Add with the same Key and multiple values will store each value under the same Key in the Dictionary. 
Obtaining the values for a Key will return the HashSet with the Values of the Key.